+++ test/imad_flutter_test.dart
@@ -1,5 +1,3 @@
-import 'package:flutter_test/flutter_test.dart';
 
-import 'package:imad_flutter/imad_flutter.dart';
 
 void main() {}